misskey: fix not fetching commentary in certain cases.
Fix a bug with not being able to fetch the commentary on certain misconfigured Misskey sites like
oekakiskey.com and nijimiss.moe.

What happened was that when we did the equivalent of these two requests:

* curl -H "Accept: text/html" https://oekakiskey.com/notes/9t6ylo4flx
* curl -H "Accept: application/ld+json" https://oekakiskey.com/notes/9t6ylo4flx

...both requests returned the same cached results, even though the Accept header should make them
return different results. This is despite the fact that the server correctly returned the `Vary:
Accept` header to ensure that these requests were cached differently. They apparently have
Cloudflare misconfigured to ignore the Accept and Vary headers and cache the results regardless.