anifty: don't lookup direct images in artist's recent posts.
The Anifty extractor had a feature where, when given a direct image URL, it would search through the
artist's recent posts to find the image and extract the post data.

This behavior was unlike what other extractors do, and it didn't always work because it only looked
at the first page of the artist's posts. It would fail once the image fell off the first page. To
make this work in general, it would have to search through all of the artist's posts, which is
infeasible.