jobs: add GoodJob admin dashboard.
Add GoodJob's builtin dashboard at /good_job. This is better than our own dashboard at /jobs, but
it's admin-only because it allows full access to jobs.

In order to make it admin-only, we have to import the user authorization and exception handling code
from ApplicationController. We also have to change the default layout so that it works when rendering
the access denied page. `main_app` is used so the routes refer to Danbooru's routes rather than
GoodJob's routes.