Fix #5952: Persist disabling individual blacklist items.
Store enabled state of individual blacklist rules in local storage.

Using local storage means that the enabled state of rules is stored per-device, not per-account.
This means that disabling a rule on your phone won't disable it on your desktop too, which may or
may not be what you want.

It also means that if someone has multiple accounts, the enabled state will be shared between all of
them on the same device, even if their accounts have totally different blacklists. This is probably
wrong, but it's how other UI state is stored too.

The previous "dab" (disable all blacklists) cookie is not respected. This means disabled blacklists
will become enabled again and will have to be redisabled. We could migrate the cookie, but it could
take months for everyone to visit the site to update the cookie and it's not worth supporting forever.