Tag autocomplete: Insert extra space after caret when using Ctrl+Backspace
If the user's caret is positioned at the start of a tag and they use Ctrl+Backspace to delete the previous tag before that, normally the caret would be left at the start of the second tag, which would interfere with autocomplete results unless the user manually repositions their caret.

By adding an extra space automatically after the user presess Ctrl+Backspace, it is possible for them to immediately start typing a new tag with no interruption.