Autocomplete: Update search results when selection changes
Currently the autocomplete search results only update when the user types something and not when they simply move the caret somewhere else in the existing text. This can cause an issue where the results shown will become outdated if the user moves their caret to a different tag, and if the user then presses tab to complete the results, it will result in the wrong tag being overwritten.

This can be fixed by manually triggering an autocomplete search when the selection changes if the autocomplete menu is already open.