css: clean up jQuery UI's css.
Get rid of all of jQuery UI's builtin CSS rules. Instead manually define only the rules we actually
need for dialog boxes and autocomplete.

This fixes a number of problems caused by the default jQuery theme applying a bunch of unwanted CSS
rules that we had to fight against to override.

* Fixes a bug where if you hovered over a colored username in autocomplete, the username would lose
  its color.
* Fixes the pin button in the popout tag edit dialog box not toggling state when clicked.
* Fixes tags being bigger in the popout tag edit dialog box than in the normal tag edit form.
* Fixes certain colors being hardcoded instead of using variables.
* Fixes some minor alignment issues in dialog boxes.
* Gets rid of a bunch of unused icon files.