dtext: improve wrapping of multi-line embedded image captions.
Use `text-wrap: balance` to make each line in a multi-line image caption have equal width. Fixes it
so that single words aren't left on a line by themselves as often.

This means that it's less often necessary to use [br] tags to force captions to wrap at the right
place. It also means that long costume tags, where you can't insert a [br] in the middle, wrap better.

https://tailwindcss.com/docs/text-wrap
https://developer.mozilla.org/en-US/docs/Web/CSS/text-wrap
https://developer.chrome.com/docs/css-ui/css-text-wrap-balance