Tag autocomplete: Only insert an extra space when deleting a word if it is the last word of a tag
Previously, Ctrl+Backspace/Delete would need to add an extra space after the caret so that any text immediately following the caret wouldn't interfere with autocomplete.

Now that the caret is treated as a wildcard in the middle of a tag, the space is no longer necessary when the word that got deleted was not the last word of a tag.
e.g. In `hand_on_own_|hip`, Ctrl+Backspace will now not insert any space, leaving you with `hand_on_|hip` so that you can start typing a replacement 3rd word immediately without the 4th word being ignored due to the space.

However, the space is still necessary in the original scenario, when your caret is positioned at the beginning of a tag and you delete the final word of the previous tag.
e.g. In `hand_on_own_hip |1girl`, Ctrl+Backspace still needs to add a space to replace the one that got deleted so that 1girl does not interfere with autocomplete, leaving you with `hand_on_own_| 1girl`.