posts: fix double unvoting bug.
Fix a race condition that allowed the score to be incremented twice when a user removed a vote. This
happened because when we didn't lock the vote while it was updated, so it was possible to soft
delete the same vote twice.

This also removes the `DELETE /posts/:post_id/vote` endpoint in favor of `DELETE /post_votes/:id`.
Using the post ID to locate the vote was also potentially racy because the same post ID could refer
to different votes if the user unvoted fast enough.

Downstream boorus should fix their scores using the fix script:

   # Do a dry run showing what will be fixed
   ./script/fixes/139_recalculate_post_scores.rb

   # Actually fix the posts
   FIX=true ./script/fixes/139_recalculate_post_scores.rb